+++ NEWS.md
@@ -1,6 +1,8 @@
 # spant 3.10.0
 * qr.solve reverts to ginv for rats in rare cases where qr.solve fails.
 * fit_svs updated to only accept a single argument for remove_basis.
+* Added cross-platform support for T1 tissue segmentation using ANTs via the
+rpyANTs package.
 
 # spant 3.9.0
 * Added extra pre_align options to fit_svs.