+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 3.10.0
 * qr.solve reverts to ginv for rats in rare cases where qr.solve fails.
 * fit_svs updated to only accept a single argument for remove_basis.
+* Bug fix for Siemens DICOM reader.
 * Added cross-platform support for T1 tissue segmentation using ANTs via the
 rpyANTs package.